src/requirements.txt
opencv-python==4.6.0.66
pytesseract==0.3.10
flask==2.2.5
moviepy==1.0.3
yt-dlp==2023.3.4
easyocr==1.7.0
numpy==1.24.1
pywebview==4.1